requirements.txt
Django>=4.1.4,<4.2
python-dotenv>=0.21.0,<0.22
Pillow>=8.3.2,<8.4
django-jazzmin>=2.6.0,<2.7
django-picklefield==3.1
django-constance>=2.9.1,<2.10
requests>=2.28.1,<2.29
